VBE WebSocket
