+++ DAVID.sln
@@ -11,6 +11,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "TEST", "TEST", "{19F54D88-6
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestProject1", "TEST\TestProject1\TestProject1.csproj", "{307E3CB0-E2EB-4FBA-B8B7-ADADE8E818F2}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DAVIDSocketServer", "DavidCore\DAVIDSocketServer\DAVIDSocketServer.csproj", "{1C2D18D6-C94F-44F7-BDB5-16E316E6976F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -28,9 +30,14 @@ Global
 		{307E3CB0-E2EB-4FBA-B8B7-ADADE8E818F2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{307E3CB0-E2EB-4FBA-B8B7-ADADE8E818F2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{307E3CB0-E2EB-4FBA-B8B7-ADADE8E818F2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1C2D18D6-C94F-44F7-BDB5-16E316E6976F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1C2D18D6-C94F-44F7-BDB5-16E316E6976F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1C2D18D6-C94F-44F7-BDB5-16E316E6976F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1C2D18D6-C94F-44F7-BDB5-16E316E6976F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{7854A452-8F43-423D-BC7E-1C1447A6B0A5} = {B1351CFC-D38B-45D1-BC6C-30995EDCDEFB}
 		{307E3CB0-E2EB-4FBA-B8B7-ADADE8E818F2} = {19F54D88-648F-4331-8AB1-AB7CDD693708}
+		{1C2D18D6-C94F-44F7-BDB5-16E316E6976F} = {B1351CFC-D38B-45D1-BC6C-30995EDCDEFB}
 	EndGlobalSection
 EndGlobal
