+++ DavidCore/DAVIDSocketServer/Class1.cs
@@ -1,0 +1,6 @@
+﻿namespace DAVID.SocketServer;
+
+public class Class1
+{
+
+}
